pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apst-starter-kit"
version = "0.1.0"
description = "Local-first Accelerated Prompt Stress Testing starter kit"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Keita Broadwater"}]
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
providers = [
    "openai>=1.0.0",
    "together>=1.0.0",
]
airbench = [
    "datasets>=2.14.0",
    "huggingface-hub>=0.16.0",
]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.4.0",
]
all = [
    "openai>=1.0.0",
    "together>=1.0.0",
    "datasets>=2.14.0",
    "huggingface-hub>=0.16.0",
    "pytest>=7.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
apst = "llm_eval.cli:main"
llm-eval = "llm_eval.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
